+++ src/index.tsx
@@ -2,82 +2,18 @@ import React from 'react'
 export { includeSubmittedSearch } from './util'
 import SearchController from './SearchController'
 import HTTPSearchHOC from './HTTPSearchHOC'
+import { Schema, Search, SearchResponse } from './types'
 
-interface Schema {
+type SearchKitSetupFunction = () => ({
+  hydratedSearch: Search
+  schemas: { [key: string]: Schema }
+})
 
-}
-
-interface Schemas {
-
-}
-
-type FacetValue = string | number
-interface Filter {
-  key: string
-  type: string
-  field: string
-  idPath?: string
-  label?: string
-  values?: FacetValue[]
-  checked?: boolean
-  from?: Date
-  to?: Date
-  interval?: string
-  isMongoId?: boolean
-  offset?: number
-  optionSearch?: string
-  include?: string[]
-  subqueryCollection?: string
-  subqueryKey?: string
-  optionsAreMongoIds?: boolean
-  subqueryField?: string
-  subqueryIdPath?: string
-  subqueryFieldIsArray?: boolean
-  subqueryLocalField?: string
-  subqueryLocalIdPath?: string
-}
-
-interface TotalsBarColumn {
-  key: string
-  field: string
-  label: string
-  agg: string
-}
-
-interface SalesTableRow {
-  key: string
-  field?: string
-  agg?: string
-  header?: boolean
-  subItem?: boolean
-  unwind?: boolean
-  hide?: boolean
-  negative?: boolean
-  
-}
-
-interface Chart {
-  key: string
-  type: string
-  currency?: string
-  hideLabel?: boolean
-  columns?: TotalsBarColumn[]
-  rows?: SalesTableRow
-}
-
-interface Search {
-  id?: string,
-  collection: string
-  filters?: Filter[]
-  pageSize?: number
-  page?: number
-  omitFromResults?: string[]
-  charts: Chart[]
-}
+type InitialSearchRunner = (search: Search) => SearchResponse
 
 interface SearchkitSetup {
-  preparePage: any
-  runInitialSearch: any
+  preparePage: SearchKitSetupFunction
+  runInitialSearch: InitialSearchRunner
   overridesStateless: any
   UIComponentsStateless: any
   FeathersSearchClientRenderer: any
